+++ frigate/__main__.py
@@ -87,7 +87,7 @@ def main() -> None:
             if current != full_config:
                 print(f"Line #  : {line_number}")
                 print(f"Key     : {' -> '.join(map(str, error_path))}")
-                print(f"Value   : {error.get('input','-')}")
+                print(f"Value   : {error.get('input', '-')}")
             print(f"Message : {error.get('msg', error.get('type', 'Unknown'))}\n")
 
         print("*************************************************************")
